refactor: 改造request

+++ src/utils/request.ts
@@ -15,7 +15,7 @@ type Result<T = any> = {
   data?: T;
 };
 
-function showConfirm(message: string, showCancelButton = false) {
+function showConfirm(message: string, showCancelButton = false, needReload = false) {
   ElMessageBox.confirm(message, "提示", {
     confirmButtonText: "确定",
     showCancelButton: false,
@@ -24,8 +24,10 @@ function showConfirm(message: string, showCancelButton = false) {
     showClose: false,
     type: "warning",
   }).then(() => {
-    localStorage.clear();
-    window.location.href = "/";
+    if (needReload) {
+      localStorage.clear();
+      window.location.href = "/";
+    } 
   });
 }
 
@@ -86,51 +88,13 @@ export default class Request {
       (error: AxiosError) => {
         // 当响应异常时做一些处理
         if (error && error.response) {
-          switch (error.response.status) {
-            case 400:
-              error.message = "请求错误(400)";
-              break;
-            case 401:
-              error.message = "授权已失效，请重新登录(401)";
-              break;
-            case 403:
-              error.message = "拒绝访问(403)";
-              break;
-            case 404:
-              error.message = "请求出错(404)";
-              break;
-            case 408:
-              error.message = "请求超时(408)";
-              break;
-            case 500:
-              error.message = "服务器错误(500)";
-              break;
-            case 501:
-              error.message = "服务未实现(501)";
-              break;
-            case 502:
-              error.message = "网络错误(502)";
-              break;
-            case 503:
-              error.message = "服务不可用(503)";
-              break;
-            case 504:
-              error.message = "网络超时(504)";
-              break;
-            case 505:
-              error.message = "HTTP版本不受支持(505)";
-              break;
-            default:
-              error.message = `连接出错(${error.response.status})!`;
+          const response: any = error.response;
+          if (response.data && response.data.msg) {
+            showConfirm(response.data.msg, false, false);
+            return Promise.reject(new Error(error.message));
           }
-          if (error.response.status < 500) {
-            showConfirm(error.message);
-          } else {
-            ElMessage.error(error.message);
-          }
-          return Promise.reject(new Error(error.message));
         }
-        showConfirm(`请求错误：${error.message}`);
+        showConfirm(`请求错误：${error.message}`, false, false);
         return Promise.reject(error.message);
       }
     );
